消息类型放到独立的文件
